Move test data to its own file
