+++ internal/resolvers/wikipedia/article_loader_test.go
@@ -2,51 +2,13 @@ package wikipedia
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
-	"net/http"
-	"net/http/httptest"
 	"net/url"
 	"testing"
 
 	"github.com/Chatterino/api/internal/logger"
-	"github.com/Chatterino/api/pkg/utils"
 	qt "github.com/frankban/quicktest"
-	"github.com/go-chi/chi/v5"
 )
-
-var (
-	wikiData = map[string]*wikipediaAPIResponse{}
-)
-
-func init() {
-	wikiData["en_test"] = &wikipediaAPIResponse{
-		Titles: wikipediaAPITitles{
-			Normalized: "Test title",
-		},
-		Extract:     "Test extract",
-		Thumbnail:   nil,
-		Description: utils.StringPtr("Test description"),
-	}
-
-	wikiData["en_test_html"] = &wikipediaAPIResponse{
-		Titles: wikipediaAPITitles{
-			Normalized: "<b>Test title</b>",
-		},
-		Extract:     "<b>Test extract</b>",
-		Thumbnail:   nil,
-		Description: utils.StringPtr("<b>Test description</b>"),
-	}
-
-	wikiData["en_test_no_description"] = &wikipediaAPIResponse{
-		Titles: wikipediaAPITitles{
-			Normalized: "Test title",
-		},
-		Extract:     "Test extract",
-		Thumbnail:   nil,
-		Description: nil,
-	}
-}
 
 func testLoadAndUnescape(ctx context.Context, loader *ArticleLoader, c *qt.C, locale, page string) (cleanTooltip string) {
 	urlString := fmt.Sprintf("https://%s.wikipedia.org/wiki/%s", locale, page)
@@ -64,25 +26,7 @@ func testLoadAndUnescape(ctx context.Context, loader *ArticleLoader, c *qt.C, lo
 func TestLoad(t *testing.T) {
 	ctx := logger.OnContext(context.Background(), logger.NewTest())
 	c := qt.New(t)
-	r := chi.NewRouter()
-	r.Get("/api/rest_v1/page/summary/{locale}/{page}", func(w http.ResponseWriter, r *http.Request) {
-		locale := chi.URLParam(r, "locale")
-		page := chi.URLParam(r, "page")
-
-		var response *wikipediaAPIResponse
-		var ok bool
-
-		if response, ok = wikiData[locale+"_"+page]; !ok {
-			http.Error(w, http.StatusText(404), 404)
-			return
-		}
-
-		b, _ := json.Marshal(&response)
-
-		w.Header().Set("Content-Type", "application/json")
-		w.Write(b)
-	})
-	ts := httptest.NewServer(r)
+	ts := testServer()
 	defer ts.Close()
 
 	loader := &ArticleLoader{
